+++ src/HasUuidPrimaryKey.php
@@ -2,20 +2,9 @@
 
 namespace Spatie\BinaryUuid;
 
+/**
+ * @deprecated The logic of this trait is moved to `Spatie\BinaryUuid\HasBinaryUuid` and this trait will be removed with the next major version.
+ */
 trait HasUuidPrimaryKey
 {
-    public function getKeyName()
-    {
-        return 'uuid';
-    }
-
-    public function getIncrementing()
-    {
-        return false;
-    }
-
-    public function resolveRouteBinding($value)
-    {
-        return $this->withUuid($value)->first();
-    }
 }